station/requirements.txt
fastapi[standard]
python-dotenv[cli]
htmx